+++ src/dft/Log3Quadrature.hpp
@@ -41,12 +41,12 @@ class CLog3Quadrature
     /**
      The number of radial points.
      */
-    int32_t _nRadialPoints;
+    int64_t _nRadialPoints;
 
     /**
      The identifier of chemical element.
      */
-    int32_t _idElemental;
+    int64_t _idElemental;
 
     /**
      Gets Xi factor of radial quadrature for specific chemical element.
@@ -62,7 +62,7 @@ class CLog3Quadrature
      @param nRadialPoints the number of radial points.
      @param idElemental the identifier of chemical element.
      */
-    CLog3Quadrature(const int32_t nRadialPoints, const int32_t idElemental);
+    CLog3Quadrature(const int64_t nRadialPoints, const int64_t idElemental);
 
     /**
      Destroys a Log3 quadrature object.
